remove unneeded temp variable
And the assert() call also: tl elements are appended
once per loop iteration, and the loop counter is equals
to acts.count(). Therefore, tl.size() <= acts.count().

Finally, RandomKnuth::pick(int n) returns an integer as a
result of X % n. It means that the returned value must be
between 0 and n-1. There is no need to assert these boundaries.

Signed-off-by: Rodolfo Ribeiro Gomes <rodolforg@gmail.com>