simplify constraint ::fitness() method parameters
The cl (constraint conflict weight) list and dl
(constraint conflict weight) list are only filled
when conflictsString is not a null pointer.

So it makes sense put them all together in a single structure,
as they work together.

In addition, conflictsString is rarely used, and it is
just a single string with all 'dl' list contents. So I removed
it, and just join those list items when required (class Solution)

Signed-off-by: Rodolfo Ribeiro Gomes <rodolforg@gmail.com>

Conflicts:
	src/engine/solution.cpp
	src/engine/spaceconstraint.cpp
	tests/engine/timeconstraints/ConstraintStudentsMinContinuousGapInInterval/tst_constraintstudentsmincontinuousgapinintervaltest.cpp
	tests/engine/timeconstraints/ConstraintTeachersMinContinuousGapInInterval/tst_constraintteachersmincontinuousgapinintervaltest.cpp